backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legaltech-backend"
version = "1.0.0"
description = "LegalTech MVP Backend API"
authors = [
    {name = "Liong Cheng Lex", email = "chenglex1@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "pdfplumber>=0.10.3",
    "spacy>=3.7.2",
    "nltk>=3.8.1",
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    "httpx>=0.25.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true